Part 5

+++ src/app/modules/tour/tour.service.ts
@@ -1,4 +1,6 @@
 import AppError from "../../errorHelpers/AppError";
+import { QueryBuilder } from "../../utils/QueryBuilder";
+import { searchConst } from "./tour.constant";
 import { ITour, ITourType } from "./tour.interface";
 import { Tour, TourType } from "./tour.model";
 import httpStatus from "http-status-codes";
@@ -10,13 +12,13 @@ const createTourTypes = async (payload: ITourType) => {
     throw new AppError(httpStatus.BAD_REQUEST, "Tour type already exists.");
   }
 
-  const tourTypes = await TourType.create({ name });
+  const tourTypes = await TourType.create(payload);
 
   return tourTypes;
 };
 
 const getAllTourTypes = async () => {
-  const allTourTypes = await TourType.find();
+  const allTourTypes = await TourType.find({});
   return allTourTypes;
 };
 
@@ -60,6 +62,81 @@ const createTour = async (payload: ITour) => {
   return tour;
 };
 
+/** OLD get all tours */
+// const getAllTours = async (query: Record<string, string>) => {
+//   const filter = query;
+//   const searchTerm = query.search || "";
+//   const sort = query.sort || "-createdAt";
+//   const fieldFiltering = query.field?.split(",").join(" ") || "";
+
+//   const page = Number(query.page) || 1;
+//   const limit = Number(query.limit) || 10;
+//   const skip = (page - 1) * limit;
+
+//   // to delete field from the filter
+//   for (const field of excludeFields) {
+//     // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+//     delete filter[field];
+//   }
+
+//   const search = {
+//     $or: searchConst.map((field) => ({
+//       [field]: { $regex: searchTerm, $options: "i" },
+//     })),
+//   };
+
+//   // const tours = await Tour.find(search)
+//   //   .find(filter)
+//   //   .sort(sort)
+//   //   .select(fieldFiltering)
+//   //   .skip(skip)
+//   //   .limit(limit);
+
+//   const filterQuery = Tour.find(filter);
+//   const tours = filterQuery.find(search);
+//   const allTours = await tours
+//     .sort(sort)
+//     .select(fieldFiltering)
+//     .skip(skip)
+//     .limit(limit);
+
+//   const totalTours = await Tour.countDocuments();
+//   const totalPage = Math.ceil(totalTours / limit);
+
+//   const meta = {
+//     total: totalTours,
+//     totalPage: totalPage,
+//     currentPage: page,
+//     limit: limit,
+//   };
+
+//   return {
+//     data: allTours,
+//     meta: meta,
+//   };
+// };
+
+const getAllTours = async (query: Record<string, string>) => {
+  const queryBuilder = new QueryBuilder(Tour.find(), query);
+
+  const tours = await queryBuilder
+    .search(searchConst)
+    .filter()
+    .sort()
+    .fields()
+    .paginate();
+
+  const [data, meta] = await Promise.all([
+    tours.build(),
+    queryBuilder.getMeta(),
+  ]);
+
+  return {
+    data,
+    meta,
+  };
+};
+
 const updateTour = async (id: string, payload: Partial<ITour>) => {
   const existingTour = await Tour.findById(id);
 
@@ -73,6 +150,12 @@ const updateTour = async (id: string, payload: Partial<ITour>) => {
 };
 
 const deleteTour = async (id: string) => {
+  const existingTour = await Tour.findOne({ _id: id });
+
+  if (!existingTour) {
+    throw new AppError(httpStatus.BAD_REQUEST, "No Tour found.");
+  }
+
   const tour = await Tour.findByIdAndDelete(id);
   return tour;
 };
@@ -83,6 +166,7 @@ export const tourService = {
   updateTourType,
   deleteTourType,
   createTour,
+  getAllTours,
   updateTour,
   deleteTour,
 };
